Separate service gen and definitions. More generation from runit.
systemd service generator now understands runit finish and conf. The
general service model aquired a few new fields for the aforementioned
things, plus a field for the healthchecking (which is not generated
for systemd, because systemd can't do healthchecks with arbitary
scripts easily). Overall, this should cover most of runit
services. log from runit is left unimplemented, as other service
daemons have their own means of logging.
